Remove deprecation warning on app boot
This version of rails added a deprecation warning
during app boot. The deprecation message was:

```
DEPRECATION WARNING: Initialization autoloaded the constant MemoryDb.

Being able to do this is deprecated. Autoloading during initialization is going
to be an error condition in future versions of Rails.

Reloading does not reboot the application, and therefore code executed during
initialization does not run again. So, if you reload MemoryDb, for example,
the expected changes won't be reflected in that stale Class object.

This autoloaded constant has been unloaded.
```

The solution was to load this file after the app was initialized.